generative-ai/multi-agents/5-practice/automating-ecommerce-agent/pyproject.toml
[project]
name = "automating-ecommerce-agent"
version = "0.1.0"
description = "Customer Support Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.25",
    "langchain-community>=0.3.12",
    "langchain-openai>=0.2.12",
    "langchain-chroma>=0.1.4",
    "langchainhub>=0.1.21",
    "langgraph>=0.2.59",
    "langsmith>=0.2.3",
    "langserve[all]>=0.3.0",
    "langchain-cli>=0.0.35",
    "streamlit>=1.41.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "google-search-results>=2.4.2",
    "st-pages>=1.0.1",
    "ipython>=8.31.0",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.2",
    "rebuff>=0.0.5",
    "onnxruntime==1.15.0",
    "supabase>=2.15.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.7.1",
    "agentevals>=0.0.8",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "remote-pdb>=2.1.0",
    "ruff>=0.8.3",
    "sqlalchemy>=2.0.36",
]

[tool.uv]
default-groups = ["dev"]

[tool.uv.workspace]
members = ["server/crawler", "server/web-crawler"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "server"
"agent" = "server"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.isort]
profile = "black"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff]
line-length = 120